Reverting changes (#5)
The new updates to the plugin commands are good changes, and would
improve the organization of the plugin, but unfortunately Aseprite's
menu API doesn't like menu groups. I haven't been able to get them to
successfully work, and this version, too, is broken. The original
version of plugin.lua functions flawlessly, and it's only 3 commands so
shouldn't bloat up the menus too much.